+++ src/main/java/com/vestie/vestie/common/config/WebConfig.java
@@ -20,7 +20,7 @@ public class WebConfig implements WebMvcConfigurer {
     public void addInterceptors(InterceptorRegistry registry) {
         registry.addInterceptor(authInterceptor)
                 .addPathPatterns("/**")
-                .excludePathPatterns("/members/login", "/members");
+                .excludePathPatterns("/members/login", "/members", "/surveys/query/**");
     }
 
     @Override
